--- app/src/main/java/woowacourse/shopping/feature/list/item/ListItem.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.feature.list.item
-
-interface ListItem
