--- app/src/main/java/woowacourse/shopping/feature/list/item/MoreListItem.kt
@@ -1,3 +1,0 @@
-package woowacourse.shopping.feature.list.item
-
-class MoreListItem : ListItem
