+++ app/src/main/java/woowacourse/shopping/feature/product/detail/ProductDetailPresenter.kt
@@ -1,8 +1,8 @@
 package woowacourse.shopping.feature.product.detail
 
 import woowacourse.shopping.data.cart.CartDbHandler
-import woowacourse.shopping.feature.model.ProductState
-import woowacourse.shopping.feature.model.mapper.toDomain
+import woowacourse.shopping.model.ProductState
+import woowacourse.shopping.model.mapper.toDomain
 
 class ProductDetailPresenter(
     private val view: ProductDetailContract.View,
