+++ app/src/main/java/woowacourse/shopping/list/adapter/RecentProductListAdapter.kt
@@ -1,11 +1,11 @@
-package woowacourse.shopping.feature.list.adapter
+package woowacourse.shopping.list.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemRecentBinding
-import woowacourse.shopping.feature.list.item.ListItem
-import woowacourse.shopping.feature.list.viewholder.RecentItemViewHolder
+import woowacourse.shopping.list.item.ListItem
+import woowacourse.shopping.list.viewholder.RecentItemViewHolder
 
 class RecentProductListAdapter(
     private var items: List<ListItem>
