+++ app/src/main/java/woowacourse/shopping/list/item/ListItem.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.list.item
+
+interface ListItem
