+++ app/src/main/java/woowacourse/shopping/list/item/MoreListItem.kt
@@ -1,0 +1,3 @@
+package woowacourse.shopping.list.item
+
+class MoreListItem : ListItem
