+++ app/src/main/java/woowacourse/shopping/list/item/ProductListItem.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.list.item
+package woowacourse.shopping.list.item
 
 class ProductListItem(
     val id: Int,
