+++ app/src/main/java/woowacourse/shopping/list/item/RecentProductListItem.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.list.item
+package woowacourse.shopping.list.item
 
 class RecentProductListItem(
     val productId: Int,
