+++ app/src/main/java/woowacourse/shopping/list/viewholder/ItemHolder.kt
@@ -1,8 +1,8 @@
-package woowacourse.shopping.feature.list.viewholder
+package woowacourse.shopping.list.viewholder
 
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewbinding.ViewBinding
-import woowacourse.shopping.feature.list.item.ListItem
+import woowacourse.shopping.list.item.ListItem
 
 abstract class ItemHolder(binding: ViewBinding) : RecyclerView.ViewHolder(binding.root) {
     abstract fun bind(listItem: ListItem, onClick: (ListItem) -> Unit)
