+++ app/src/main/java/woowacourse/shopping/list/viewholder/ProductMoreViewHolder.kt
@@ -1,8 +1,8 @@
-package woowacourse.shopping.feature.list.viewholder
+package woowacourse.shopping.list.viewholder
 
 import androidx.viewbinding.ViewBinding
 import woowacourse.shopping.databinding.ItemMoreBinding
-import woowacourse.shopping.feature.list.item.ListItem
+import woowacourse.shopping.list.item.ListItem
 
 class ProductMoreViewHolder(binding: ViewBinding) : ItemHolder(binding) {
     private val binding = binding as ItemMoreBinding
