+++ app/src/main/java/woowacourse/shopping/list/viewholder/RecentListItemViewHolder.kt
@@ -1,9 +1,9 @@
-package woowacourse.shopping.feature.list.viewholder
+package woowacourse.shopping.list.viewholder
 
 import androidx.viewbinding.ViewBinding
 import woowacourse.shopping.databinding.ItemRecentProductListBinding
-import woowacourse.shopping.feature.list.adapter.RecentProductListAdapter
-import woowacourse.shopping.feature.list.item.ListItem
+import woowacourse.shopping.list.adapter.RecentProductListAdapter
+import woowacourse.shopping.list.item.ListItem
 
 class RecentListItemViewHolder(binding: ViewBinding, val items: List<ListItem>) : ItemHolder(binding) {
     private val binding = binding as ItemRecentProductListBinding
