+++ app/src/main/java/woowacourse/shopping/model/ProductState.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.model
+package woowacourse.shopping.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
