+++ app/src/main/java/woowacourse/shopping/util/extension/ContextExt.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.extension
+package woowacourse.shopping.util.extension
 
 import android.content.Context
 import android.widget.Toast
